+++ src/main/java/com/stc21/boot/auction/entity/Lot.java
@@ -55,16 +55,16 @@ public class Lot {
     private City city;
 
     @Column(name = "current_price")
-    private Double currentPrice;
+    private Long currentPrice;
 
     @Column(name = "max_price")
-    private Double maxPrice;
+    private Long maxPrice;
 
     @Column(name = "min_price")
-    private Double minPrice;
+    private Long minPrice;
 
     @Column(name = "step_price")
-    private Double stepPrice;
+    private Long stepPrice;
 
     @Column(columnDefinition = "boolean default false", nullable = false)
     private Boolean deleted;
@@ -73,4 +73,5 @@ public class Lot {
             cascade = CascadeType.ALL,
             orphanRemoval = true)
     private List<Photo> photos;
+
 }
